sticky shared setting
