+++ apps/spruce/src/pages/adminSettings/Header.tsx
@@ -1,7 +1,6 @@
 import { useRef, useState } from "react";
-import styled from "@emotion/styled";
-import { H2, H2Props } from "@leafygreen-ui/typography";
-import { size, transitionDuration, zIndex } from "@evg-ui/lib/constants/tokens";
+import { H2 } from "@leafygreen-ui/typography";
+import { stickyHeaderContainer } from "components/Settings/sharedStyles";
 import { AdminSettingsTabRoutes } from "constants/routes";
 import { AdminSettings } from "gql/generated/types";
 import useIntersectionObserver from "hooks/useIntersectionObserver";
@@ -23,35 +22,15 @@ export const Header: React.FC<Props> = ({ adminSettingsData, tab }) => {
     setShowShadow(!entry.isIntersecting);
   });
 
+  const Container = stickyHeaderContainer;
+
   return (
     <>
       <div ref={headerScrollRef} />
       <Container saveable={saveable} showShadow={showShadow}>
-        <StyledH2 data-cy="admin-settings-tab-title">{title}</StyledH2>
+        <H2 data-cy="admin-settings-tab-title">{title}</H2>
         {saveable && <AdminSaveButton adminSettingsData={adminSettingsData} />}
       </Container>
     </>
   );
 };
-
-const StyledH2 = styled(H2)<H2Props>``;
-
-const Container = styled.div<{ showShadow: boolean; saveable: boolean }>`
-  align-items: start;
-  background-color: white;
-  display: flex;
-  gap: ${size.s};
-  justify-content: space-between;
-  margin: 0 -${size.l};
-  padding: 0 ${size.l} ${size.s} ${size.l};
-
-  ${({ saveable }) => saveable && "position: sticky;"}
-  z-index: ${zIndex.stickyHeader};
-  top: 0;
-
-  ${({ showShadow }) =>
-    showShadow
-      ? "box-shadow: 0 3px 4px -4px rgba(0, 0, 0, 0.6);"
-      : "box-shadow: unset;"}
-  transition: box-shadow ${transitionDuration.default}ms ease-in-out;
-`;
