+++ apps/spruce/src/pages/projectAndRepoSettings/shared/Header.tsx
@@ -1,7 +1,7 @@
 import { useRef, useState } from "react";
 import styled from "@emotion/styled";
-import { H2 } from "@leafygreen-ui/typography";
-import { size, transitionDuration, zIndex } from "@evg-ui/lib/constants/tokens";
+import { H2, H2Props } from "@leafygreen-ui/typography";
+import { stickyHeaderContainer } from "components/Settings/sharedStyles";
 import { ProjectSettingsTabRoutes } from "constants/routes";
 import useIntersectionObserver from "hooks/useIntersectionObserver";
 import { getTabTitle } from "./getTabTitle";
@@ -30,11 +30,13 @@ export const Header: React.FC<Props> = ({ id, projectType, tab }) => {
     setShowShadow(!entry.isIntersecting);
   });
 
+  const Container = stickyHeaderContainer;
+
   return (
     <>
       <div ref={headerScrollRef} />
       <Container saveable={saveable} showShadow={showShadow}>
-        <H2 data-cy="project-settings-tab-title">{title}</H2>
+        <StyledH2 data-cy="project-settings-tab-title">{title}</StyledH2>
         {saveable && (
           <HeaderButtons
             id={id}
@@ -47,22 +49,4 @@ export const Header: React.FC<Props> = ({ id, projectType, tab }) => {
   );
 };
 
-const Container = styled.div<{ showShadow: boolean; saveable: boolean }>`
-  align-items: start;
-  background-color: white;
-  display: flex;
-  gap: ${size.s};
-  justify-content: space-between;
-  margin: 0 -${size.l};
-  padding: 0 ${size.l} ${size.s} ${size.l};
-
-  ${({ saveable }) => saveable && "position: sticky;"}
-  z-index: ${zIndex.stickyHeader};
-  top: 0;
-
-  ${({ showShadow }) =>
-    showShadow
-      ? "box-shadow: 0 3px 4px -4px rgba(0, 0, 0, 0.6);"
-      : "box-shadow: unset;"}
-  transition: box-shadow ${transitionDuration.default}ms ease-in-out;
-`;
+const StyledH2 = styled(H2)<H2Props>``;
